fixup! feat(state): add DhtStore cache mirror methods in dht_store/cache.rs
