+++ crates/holochain_state/src/source_chain.rs
@@ -499,7 +499,7 @@ impl SourceChain {
                     let mut inserted_action_hashes = std::collections::HashSet::<ActionHash>::new();
 
                     for sah in &actions_for_new_db {
-                        let new_sah = legacy_to_dht_v2_signed_action(sah);
+                        let new_sah = holochain_zome_types::dht_v2::from_legacy_signed_action(sah);
 
                         tx.insert_action(
                             &new_sah,
@@ -1503,7 +1503,7 @@ pub async fn genesis(
             &agent_action_for_new_db,
         ];
         for sah in genesis_actions {
-            let new_sah = legacy_to_dht_v2_signed_action(sah);
+            let new_sah = holochain_zome_types::dht_v2::from_legacy_signed_action(sah);
             tx.insert_action(
                 &new_sah,
                 Some(holochain_zome_types::dht_v2::RecordValidity::Accepted),
@@ -1717,79 +1717,6 @@ pub async fn dump_state(
 // ---------------------------------------------------------------------------
 // Private helpers for the new-DB writes in `flush` and `genesis`
 // ---------------------------------------------------------------------------
-
-/// Convert a legacy [`SignedActionHashed`] (using the variant-per-type `Action`
-/// enum) to the new [`holochain_zome_types::dht_v2::SignedActionHashed`] (which
-/// uses a flat `ActionHeader` + `ActionData` envelope).
-///
-/// All legacy action variants are covered. The hash is carried over from the
-/// original (the v2 hash is content-derived, so re-hashing would change it;
-/// the pre-hashed constructor preserves the existing hash as the canonical
-/// identity during the dual-write transition).
-pub(crate) fn legacy_to_dht_v2_signed_action(
-    shh: &SignedActionHashed,
-) -> holochain_zome_types::dht_v2::SignedActionHashed {
-    use holochain_zome_types::dht_v2::{
-        Action as V2Action, ActionData, ActionHeader, AgentValidationPkgData, CloseChainData,
-        CreateData, CreateLinkData, DeleteData, DeleteLinkData, DnaData, InitZomesCompleteData,
-        OpenChainData, UpdateData,
-    };
-
-    // `Action` (legacy) and `SignedHashed` are in scope via the prelude.
-    let legacy_action = shh.action();
-    let header = ActionHeader {
-        author: legacy_action.author().clone(),
-        timestamp: legacy_action.timestamp(),
-        action_seq: legacy_action.action_seq(),
-        prev_action: legacy_action.prev_action().cloned(),
-    };
-
-    let data = match legacy_action {
-        Action::Dna(d) => ActionData::Dna(DnaData {
-            dna_hash: d.hash.clone(),
-        }),
-        Action::AgentValidationPkg(d) => ActionData::AgentValidationPkg(AgentValidationPkgData {
-            membrane_proof: d.membrane_proof.clone(),
-        }),
-        Action::InitZomesComplete(_) => ActionData::InitZomesComplete(InitZomesCompleteData {}),
-        Action::Create(d) => ActionData::Create(CreateData {
-            entry_type: d.entry_type.clone(),
-            entry_hash: d.entry_hash.clone(),
-        }),
-        Action::Update(d) => ActionData::Update(UpdateData {
-            original_action_address: d.original_action_address.clone(),
-            original_entry_address: d.original_entry_address.clone(),
-            entry_type: d.entry_type.clone(),
-            entry_hash: d.entry_hash.clone(),
-        }),
-        Action::Delete(d) => ActionData::Delete(DeleteData {
-            deletes_address: d.deletes_address.clone(),
-            deletes_entry_address: d.deletes_entry_address.clone(),
-        }),
-        Action::CreateLink(d) => ActionData::CreateLink(CreateLinkData {
-            base_address: d.base_address.clone(),
-            target_address: d.target_address.clone(),
-            zome_index: d.zome_index,
-            link_type: d.link_type,
-            tag: d.tag.clone(),
-        }),
-        Action::DeleteLink(d) => ActionData::DeleteLink(DeleteLinkData {
-            base_address: d.base_address.clone(),
-            link_add_address: d.link_add_address.clone(),
-        }),
-        Action::CloseChain(d) => ActionData::CloseChain(CloseChainData {
-            new_target: d.new_target.clone(),
-        }),
-        Action::OpenChain(d) => ActionData::OpenChain(OpenChainData {
-            prev_target: d.prev_target.clone(),
-            close_hash: d.close_hash.clone(),
-        }),
-    };
-
-    let v2_action = V2Action { header, data };
-    let hashed = holo_hash::HoloHashed::with_pre_hashed(v2_action, shh.as_hash().clone());
-    SignedHashed::with_presigned(hashed, shh.signature().clone())
-}
 
 /// Return the `(cap_access_i64, Option<tag>)` parameters needed for
 /// `TxWrite::insert_cap_grant`, if the given action creates/updates a
